🐛fix: Tauri AppImage failing to render on wayland + mesa
AppImages built with Tauri fail to render with EGL_BAD_PARAMETER on
wayland + mesa. This is a side effect of binary stripping in older
linuxdeploy builds. Tauri cli pulls in an outdated build of linuxdeploy.
This can be fixed by opting out of binary stripping, or by upgrading
linuxdeploy. Tauri cli is well behaved in that it doesn't fetch
linuxdeploy if it already exists in the cache. This behavior allows us
to pin linuxdeploy to any version of our choosing by prepopulating the
toolsdirectory with a linuxdeploy binary before calling Tauri cli.