Maintain compat between ik/llama.cpp server args
Thanks louiehelm for bringing this to my attention and starting the
conversation and patch set. This approach seems the fastest way to
provide Jan users access to high quality ik_llama.cpp quantizations and
maintaining existing compatibility with mainline llama.cpp.

There are a few small differences in the exact llama-server arguments
between forks and this patch set enables support across both forks. This works
with the existing ability to install a backend from a file to use
Thireus' windows builds of ik_llama.cpp directly.